Contact
Fixed the contact page so it no longer need use client to render the basics. Also changed the poly component used in the poly blog post. Looks more inline with what I wanted.
