+++ Portfolio/src/views/scriptorium/scriptorium.jsx
@@ -1,114 +1,28 @@
-"use client";
-import { useEffect } from "react";
-import { createAnimatable, utils } from "animejs";
-
 import {
   desParticle,
   desMappingFullstack,
   desPolySVG,
 } from "./articles/articleDes.js";
 import { ScriptCard } from "../../components/scriptorium/scriptCard.jsx";
-import { AddMember, RemoveMember } from "../../utils/aniFrame.jsx";
-import { useRouter } from "next/navigation.js";
+import PillAnimation from "./pillAnimation.jsx";
 
 import "./scriptorium.css";
 
 function Scriptorium() {
-  const router = useRouter();
-  function onMouseMove(event, animatablePills, bluePills) {
-    const { clientX, clientY } = event;
-    bluePills.forEach((pill, index) => {
-      const bounding = pill.getBoundingClientRect();
-      const { left, top, width, height } = bounding;
-      // Calculate distance from the mouse to the center of the pill
-      const dx = clientX - (left + width / 2);
-      const dy = clientY - (top + height / 2);
-      const distance = Math.sqrt(dx * dx + dy * dy);
-
-      // Activate on distance
-      if (distance < 150) {
-        const hw = width / 2;
-        const hh = height / 2;
-        const x = utils.clamp(-(clientX - left - hw), -hw, hw);
-        const y = utils.clamp(-(clientY - top - hh), -hh, hh);
-        //Update x and y
-        animatablePills[index].x(x);
-        animatablePills[index].y(y);
-        pill.classList.add("hovered");
-      } else {
-        animatablePills[index].x(0);
-        animatablePills[index].y(0);
-        pill.classList.remove("hovered");
-      }
-    });
-  }
-
-  function initalize() {
-    // Selects them all
-    const bluePills = Array.from(
-      document.querySelectorAll(".script-article-tags span")
-    );
-    // Create individual animations for each BluePill
-    const animatablePills = bluePills.map((pill) =>
-      createAnimatable(pill, {
-        x: 400,
-        y: 400,
-        ease: "ease(3)",
-      })
-    );
-    return { animatablePills, bluePills };
-  }
-
-  useEffect(() => {
-    const InitVals = initalize();
-    let mouseX = 0;
-    let mouseY = 0;
-    let prevMouseX = null;
-    let prevMouseY = null;
-
-    const handleMouseMove = (event) => {
-      mouseX = event.clientX;
-      mouseY = event.clientY;
-    };
-
-    function update() {
-      // Only update if the mouse position has changed
-      if (mouseX !== prevMouseX || mouseY !== prevMouseY) {
-        onMouseMove(
-          { clientX: mouseX, clientY: mouseY },
-          InitVals.animatablePills,
-          InitVals.bluePills
-        );
-        prevMouseX = mouseX;
-        prevMouseY = mouseY;
-      }
-    }
-
-    AddMember(update);
-    window.addEventListener("mousemove", handleMouseMove);
-
-    return () => {
-      window.removeEventListener("mousemove", handleMouseMove);
-      RemoveMember(update);
-    };
-  }, []);
-
   return (
     <div className="main-scriptorium-container">
+      <PillAnimation />
       <div className="article-card-container">
-        <ScriptCard
-          articleDes={desParticle}
-          onClick={() => router.push("/scriptorium/particle")}
-        />
+        <ScriptCard articleDes={desParticle} link={"/scriptorium/particle"} />
 
         <ScriptCard
           articleDes={desMappingFullstack}
-          onClick={() => router.push("/scriptorium/mapping-fullstack")}
+          link={"/scriptorium/mapping-fullstack"}
         />
 
         <ScriptCard
           articleDes={desPolySVG}
-          onClick={() => router.push("/scriptorium/poly-svg-background")}
+          link={"/scriptorium/poly-svg-background"}
         />
       </div>
     </div>
